refactor(body-line-break-punctuation): 句読点改行ルールを整理
- モジュール分割
- `defaultTerminator`がどうせ絶対`anchorAtEnd`を通るので最初からそういう正規表現リテラルにする
